separate mlpdecoder to kale.predict
