+++ kale/embed/seq_nn.py
@@ -2,13 +2,12 @@
 DeepDTA based models for drug-target interaction prediction problem.
 """
 
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch_geometric.nn import GCNConv, global_max_pool
 
 
-class DeepDTAEncoder(nn.Module):
+class CNNEncoder(nn.Module):
     r"""
     The DeepDTA's CNN encoder module, which comprises three 1D-convolutional layers and one max-pooling layer. The module
     is applied to encoding drug/target sequence information, and the input should be transformed information with
@@ -24,7 +23,7 @@ class DeepDTAEncoder(nn.Module):
     """
 
     def __init__(self, num_embeddings, embedding_dim, sequence_length, num_kernels, kernel_length):
-        super(DeepDTAEncoder, self).__init__()
+        super(CNNEncoder, self).__init__()
         self.embedding = nn.Embedding(num_embeddings + 1, embedding_dim)
         self.conv1 = nn.Conv1d(in_channels=sequence_length, out_channels=num_kernels, kernel_size=kernel_length)
         self.conv2 = nn.Conv1d(in_channels=num_kernels, out_channels=num_kernels * 2, kernel_size=kernel_length)
@@ -41,7 +40,7 @@ class DeepDTAEncoder(nn.Module):
         return x
 
 
-class DrugGCNEncoder(nn.Module):
+class GCNEncoder(nn.Module):
     r"""
     The GraphDTA's GCN encoder module, which comprises three graph convolutional layers and one full connected layer.
     The model is a variant of DeepDTA and is applied to encoding drug molecule graph information. The original paper
@@ -54,7 +53,7 @@ class DrugGCNEncoder(nn.Module):
         dropout_rate (float): dropout rate during training.
     """
     def __init__(self, in_channel=78, out_channel=128, dropout_rate=0.2):
-        super(DrugGCNEncoder, self).__init__()
+        super(GCNEncoder, self).__init__()
         self.conv1 = GCNConv(in_channel, in_channel)
         self.conv2 = GCNConv(in_channel, in_channel * 2)
         self.conv3 = GCNConv(in_channel * 2, in_channel * 4)
@@ -69,34 +68,4 @@ class DrugGCNEncoder(nn.Module):
         x = global_max_pool(x, batch)
         x = self.fc(x)
         x = self.dropout(x)
-        return x
-
-
-class MLPDecoder(nn.Module):
-    r"""
-    The MLP decoder module, which comprises four fully connected neural networks. It's a common decoder for decoding
-    drug-target encoding information.
-
-    Args:
-        in_dim (int): Dimension of input feature.
-        hidden_dim (int): Dimension of hidden layers.
-        out_dim (int): Dimension of output layer.
-        dropout_rate (float): dropout rate during training.
-    """
-    def __init__(self, in_dim, hidden_dim, out_dim, dropout_rate=0.1):
-        super(MLPDecoder, self).__init__()
-        self.fc1 = nn.Linear(in_dim, hidden_dim)
-        self.fc2 = nn.Linear(hidden_dim, hidden_dim)
-        self.fc3 = nn.Linear(hidden_dim, out_dim)
-        self.fc4 = nn.Linear(out_dim, 1)
-        torch.nn.init.normal_(self.fc4.weight)
-        self.dropout = nn.Dropout(dropout_rate)
-
-    def forward(self, x):
-        x = F.relu(self.fc1(x))
-        x = self.dropout(x)
-        x = F.relu(self.fc2(x))
-        x = self.dropout(x)
-        x = F.relu(self.fc3(x))
-        x = self.fc4(x)
         return x
